Uploaded files from desktop computations
Uploaded saturation W-dependence tests for W = [10^-2, 10^4], evenly spaced configurations for N=4, and at varying  (rings x rays) domain sizes between 8x8 to 64x64.  There seems to be a numerical instability for the 8x8 computation, since W=10^4 results in a significantly small MFPT value relative to prior entries in the table. However, this is issue is nothing new, and is byproduct of the current delta-time computation scheme. For smaller domain sizes, multiplying delta-time by an additional factor of 0.1 prevents the instability from occurring.