Collection of mass analysis results.
The following commit is comprised of numerical and graphical results of mass as a function of time on the diffusive and advective layer. In addition, there are results for the fraction of particles active on filaments; the mass on the advective layer divided by the total mass in the domain.